pyproject.toml
[tool.poetry]
name = "final-year-project-2024"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pymongo = {extras = ["srv"], version = "^4.10.1"}
pandas = "^2.2.3"
numpy = "^2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
